Added Assertions: todo_ event_ deadline_ find_ done_ done_ _

+++ src/main/java/DeleteCommand.java
@@ -27,18 +27,29 @@ public class DeleteCommand extends Command {
     public String execute(TaskList task, Ui ui, Storage storage) throws DukeException {
         String output = "";
         String[] deletearr = command.split(" ");
-        int number = Integer.parseInt(deletearr[1]) - 1;
-        Task deletetask = task.getTaskList().get(number);
-        output += ("     Noted. I've removed this task:\n");
-        output += ("       " + deletetask.toString() + "\n");
-        task.deleteTask(number);
-        output += ("     Now you have " + task.getCount() + " tasks in the list.\n");
         try {
-            storage.deleteWrite(number);
-        } catch (IOException e) {
-            System.out.println(e);
+            assert deletearr.length > 1 : " Please input the task number you want to delete";
+            assert deletearr.length < 3 : " Please only input a single number for the task you want to delete";
+            if (deletearr.length == 1 || deletearr.length > 2) {
+                throw new DukeException(" ☹ OOPS!!! Please input the task that you want to delete.");
+            }
+            int number = Integer.parseInt(deletearr[1]) - 1;
+            Task deletetask = task.getTaskList().get(number);
+            output += ("     Noted. I've removed this task:\n");
+            output += ("       " + deletetask.toString() + "\n");
+            task.deleteTask(number);
+            output += ("     Now you have " + task.getCount() + " tasks in the list.\n");
+            try {
+                storage.deleteWrite(number);
+            } catch (IOException e) {
+                System.out.println(e);
+            }
+            return output;
+        } catch (AssertionError e) {
+            return e.toString();
+        } catch (DukeException e) {
+            return e.toString();
         }
-        return output;
     }
 
 }
